% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auth_basic}
\alias{auth_basic}
\title{Generate Basic authentication strings}
\usage{
auth_basic(user, password)
}
\arguments{
\item{user}{A vector of usernames.}

\item{password}{A vector of passwords.}
}
\value{
A character vector of encoded credentials.
}
\description{
Combine usernames with passwords with colons, and generate base64-encoded
strings to be used for user authentication.
}
\examples{
servr::auth_basic("foo", "B@R")
}
