% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daemon_stop}
\alias{daemon_stop}
\alias{daemon_list}
\title{Utilities for daemonized servers}
\usage{
daemon_stop(which = daemon_list())

daemon_list()
}
\arguments{
\item{which}{A integer vector of the server IDs; by default, IDs of all
existing servers in the current R session obtained from
\code{daemon_list()}, i.e., all daemon servers will be stopped by default.}
}
\value{
The function \code{daemon_list()} returns a list of existing server
  IDs, and \code{daemon_stop()} returns an invisible \code{NULL}.
}
\description{
\code{daemon_list()} returns IDs of servers, which can be used to stop the
daemonized servers.
}
