% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebook}
\alias{notebook}
\title{Serve R notebooks}
\usage{
notebook(dir = ".", ...)
}
\arguments{
\item{dir}{The root directory to serve.}

\item{...}{server configurations passed to \code{\link{server_config}()}}
}
\description{
Serve R notebook files created from the \pkg{rnotebook} package.
}
\details{
An R notebook is basically a JSON file, and the JSON data is inserted to a
predefined template in this package (\code{system.file('resources',
'rnotebook.html', package = 'servr')}), and rendered through the JavaScript
implementation of \href{https://commonmark.org}{CommonMark} at the moment (the
renderer may change in the future).
}
\examples{
# see ?rnotebook::newnb for examples
}
\keyword{internal}
