% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_port}
\alias{random_port}
\title{Find a random available TCP port}
\usage{
random_port(port = 4321L, host = getOption("servr.host", "127.0.0.1"), n = 20)
}
\arguments{
\item{port}{The preferred port(s).}

\item{host}{A string that is a valid IPv4 address that is owned by this
server, or \code{"0.0.0.0"} to listen on all IP addresses.}

\item{n}{The maximum number of random ports to be tested.}
}
\value{
A port number, or an error if no ports are available.
}
\description{
Test a series of random TCP ports from 3000 to 8000 (excluding a few that are
considered unsafe by Chrome) and return the first available one. A web server
can be later started on this port.
}
