% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect.R
\name{redirect}
\alias{redirect}
\title{Create a redirect response}
\usage{
redirect(dest, status = 301L)
}
\arguments{
\item{dest}{A destination path.}

\item{status}{The status code (usually \code{301} or \code{302}).}
}
\description{
Create a response to redirect to a destination.
}
\examples{
servr::redirect("https://www.r-project.org")
}
