% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect.R
\name{redirect}
\alias{redirect}
\title{Create a redirect server}
\usage{
redirect(dest, ...)
}
\arguments{
\item{dest}{A function that takes the requested path as input and returns a
destination path.}

\item{...}{Arguments to be passed to \code{\link{server_config}()}.}
}
\description{
Create a server that simply redirect a requested path to another path.
}
\examples{\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# redirect any request to httpbin.org
s = servr::redirect(dest = function(path) {
    paste0("https://httpbin.org/anything?path=", path)
})
s$url

browseURL(paste0(s$url, "/hello"))
browseURL(paste0(s$url, "/world"))

s$stop_server()
\dontshow{\}) # examplesIf}
}
